requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pytest
pytest-django
pytest-cov
PyYAML
flake8
python-dotenv

